Print out the exception details when we can't start a debug session
I had a bug in XS where clicking the 'debug' button pretty much
immediately aborted the operation. There was no output because
it was being swallowed here. I was told this issue had occurred
before and we couldn't figure it out, so I'd like to add logging
here so we get some output when things explode in this unexpected
way.